test(evidence): share visual verification primitives
